snippingtool/pyproject.toml
[project]
name = "snippingtool"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai>=1.78.1",
    "pillow>=11.2.1",
    "pynput>=1.8.1",
    "python-dotenv>=1.1.0",
]
